Move plantuml to diagrams pkg
