+++ pkg/diagrams/plantuml.go
@@ -1,9 +1,7 @@
-package main
+package diagrams
 
 import (
 	"os"
-
-	"github.com/anz-bank/sysl/pkg/diagrams"
 
 	"github.com/spf13/afero"
 
@@ -15,11 +13,11 @@ const (
 	PlantUMLDefault = "http://localhost:8080/plantuml"
 )
 
-type plantumlmixin struct {
+type Plantumlmixin struct {
 	value string
 }
 
-func (p *plantumlmixin) AddFlag(cmd *kingpin.CmdClause) {
+func (p *Plantumlmixin) AddFlag(cmd *kingpin.CmdClause) {
 	cmd.Flag("plantuml",
 		"base url of plantuml server (default: "+PlantUMLEnvVar+" or "+
 			PlantUMLDefault+" see "+
@@ -27,7 +25,7 @@ func (p *plantumlmixin) AddFlag(cmd *kingpin.CmdClause) {
 	).Short('p').StringVar(&p.value)
 }
 
-func (p *plantumlmixin) Value() string {
+func (p *Plantumlmixin) Value() string {
 	if p.value == "" {
 		p.value = os.Getenv(PlantUMLEnvVar)
 		if p.value == "" {
@@ -37,9 +35,9 @@ func (p *plantumlmixin) Value() string {
 	return p.value
 }
 
-func (p *plantumlmixin) GenerateFromMap(m map[string]string, fs afero.Fs) error {
+func (p *Plantumlmixin) GenerateFromMap(m map[string]string, fs afero.Fs) error {
 	for k, v := range m {
-		if err := diagrams.OutputPlantuml(k, p.Value(), v, fs); err != nil {
+		if err := OutputPlantuml(k, p.Value(), v, fs); err != nil {
 			return err
 		}
 	}
